Use correct parameter names with Transform Service and also convert to TransformationOptionLimits
