+++ src/main/resources/alfresco/rendition-services2-context.xml
@@ -80,6 +80,15 @@
         <property name="nodeService" ref="NodeService" />
         <property name="contentService" ref="ContentService" />
         <property name="renditionService2" ref="renditionService2" />
+        <property name="converter" ref="transformOptionsConverter" />
+    </bean>
+
+    <!-- Deprecated bean that converts transform properties to the deprecated TransformationOptions format -->
+    <bean id="transformOptionsConverter" class="org.alfresco.repo.rendition2.TransformationOptionsConverter">
+        <property name="readLimitTimeMs" value="${system.thumbnail.definition.default.readLimitTimeMs}" />
+        <property name="readLimitKBytes" value="${system.thumbnail.definition.default.readLimitKBytes}" />
+        <property name="pageLimit" value="${system.thumbnail.definition.default.pageLimit}" />
+        <property name="maxPages" value="${system.thumbnail.definition.default.maxPages}" />
     </bean>
 
     <!-- Replaced in the enterprise edition -->
@@ -87,6 +96,7 @@
 
     <bean id="localTransformServiceRegistry" class="org.alfresco.repo.rendition2.LocalTransformServiceRegistry" >
         <property name="contentService" ref="contentService" />
+        <property name="converter" ref="transformOptionsConverter" />
     </bean>
 
     <bean id="RenditionService2_transaction"
@@ -116,12 +126,8 @@
         <constructor-arg name="registry" ref="renditionDefinitionRegistry2"/>
         <constructor-arg name="transformOptions">
             <map>
-                <entry key="timeoutMs" value="${system.thumbnail.definition.default.timeoutMs}" />
-                <entry key="readLimitTimeMs" value="${system.thumbnail.definition.default.readLimitTimeMs}" />
+                <entry key="timeout" value="${system.thumbnail.definition.default.timeoutMs}" />
                 <entry key="maxSourceSizeKBytes" value="${system.thumbnail.definition.default.maxSourceSizeKBytes}" />
-                <entry key="readLimitKBytes" value="${system.thumbnail.definition.default.readLimitKBytes}" />
-                <entry key="pageLimit" value="${system.thumbnail.definition.default.pageLimit}" />
-                <entry key="maxPages" value="${system.thumbnail.definition.default.maxPages}" />
             </map>
         </constructor-arg>
     </bean>
@@ -140,8 +146,8 @@
         <constructor-arg name="targetMimetype" value="image/jpeg"/>
         <constructor-arg name="transformOptions">
             <map merge="true">
-                <entry key="width" value="100"/>
-                <entry key="height" value="100"/>
+                <entry key="resizeWidth" value="100"/>
+                <entry key="resizeHeight" value="100"/>
             </map>
         </constructor-arg>
     </bean>
@@ -151,8 +157,8 @@
         <constructor-arg name="targetMimetype" value="image/png"/>
         <constructor-arg name="transformOptions">
             <map merge="true">
-                <entry key="width" value="100"/>
-                <entry key="height" value="100"/>
+                <entry key="resizeWidth" value="100"/>
+                <entry key="resizeHeight" value="100"/>
                 <entry key="allowEnlargement" value="false" />
             </map>
         </constructor-arg>
@@ -163,8 +169,8 @@
         <constructor-arg name="targetMimetype" value="image/jpeg"/>
         <constructor-arg name="transformOptions">
             <map merge="true">
-                <entry key="width" value="960"/>
-                <entry key="height" value="960"/>
+                <entry key="resizeWidth" value="960"/>
+                <entry key="resizeHeight" value="960"/>
                 <entry key="allowEnlargement" value="false" />
             </map>
         </constructor-arg>
@@ -175,8 +181,8 @@
         <constructor-arg name="targetMimetype" value="image/png"/>
         <constructor-arg name="transformOptions">
             <map merge="true">
-                <entry key="width" value="64"/>
-                <entry key="height" value="64"/>
+                <entry key="resizeWidth" value="64"/>
+                <entry key="resizeHeight" value="64"/>
                 <entry key="allowEnlargement" value="false" />
             </map>
         </constructor-arg>
@@ -187,8 +193,8 @@
         <constructor-arg name="targetMimetype" value="image/png"/>
         <constructor-arg name="transformOptions">
             <map merge="true">
-                <entry key="width" value="32"/>
-                <entry key="height" value="32"/>
+                <entry key="resizeWidth" value="32"/>
+                <entry key="resizeHeight" value="32"/>
                 <entry key="allowEnlargement" value="false" />
             </map>
         </constructor-arg>
