+++ src/test/resources/log4j.properties
@@ -218,7 +218,6 @@ log4j.logger.org.alfresco.repo.web.scripts.AlfrescoRhinoScriptDebugger=off
 log4j.logger.org.alfresco.repo.jscript=error
 log4j.logger.org.alfresco.repo.jscript.ScriptLogger=warn
 log4j.logger.org.alfresco.repo.cmis.rest.CMISTest=info
-
 log4j.logger.org.alfresco.repo.domain.schema.script.ScriptBundleExecutorImpl=off
 log4j.logger.org.alfresco.repo.domain.schema.script.ScriptExecutorImpl=info
 
@@ -260,5 +259,6 @@ log4j.logger.org.alfresco.repo.security.authentication.identityservice=debug
 log4j.logger.org.keycloak=debug
 
 log4j.logger.org.alfresco.repo.content.transform.TransformerDebug=debug
-
+log4j.logger.org.alfresco.repo.rendition2=debug
+log4j.logger.org.alfresco.repo.rendition.RenditionServiceImpl=debug
 log4j.logger.org.alfresco.repo.rawevents=debug
